requirements.txt
openai
instructor
pydantic
python-dotenv
tenacity
aiohttp